+++ frontend/src/components/atoms/index.ts
@@ -4,3 +4,5 @@ export { Nav } from './headers/Nav'
 export { BodyCell } from './tables/BodyCell'
 export { HeaderCell } from './tables/HeaderCell'
 export { InputCell } from './tables/InputCell'
+export { TableWrapper } from './tables/TableWrapper'
+export { MenuBtn } from './headers/MenuBtn'
