+++ frontend/src/components/molecules/index.ts
@@ -2,3 +2,4 @@ export { HeaderMole } from './HeaderMole'
 export { InputForm } from './InputForm'
 export { TableBodyMole } from './TableBodyMole'
 export { TableHeaderMole } from './TableHeaderMole'
+export { Side } from './Side'
