+++ frontend/src/components/organisms/index.ts
@@ -1,3 +1,7 @@
 export { FormOrga } from './FormOrga'
 export { HeaderOrga } from './HeaderOrga'
 export { CommonTable } from './CommonTable'
+export { HomeTable } from './HomeTable'
+export { ServerTable } from './ServerTable'
+export { PwCell } from './PwCell'
+export { SideBar } from './SideBar'
