+++ frontend/src/components/templates/index.ts
@@ -1,3 +1,4 @@
 export { EditTemp } from './EditTemp'
 export { HomeTemp } from './HomeTemp'
 export { NewTemp } from './NewTemp'
+export { ServerTemp } from './ServerTemp'
