+++ frontend/src/pages/server.tsx
@@ -1,5 +1,4 @@
-import { HomeTemp } from 'components/templates'
-import { ServerTemp } from 'components/templates/ServerTemp'
+import { ServerTemp } from 'components/templates'
 import { GetServers } from 'types/server'
 import { ServerCellDataContext } from 'utils'
 import { checkSession } from 'utils/checkSession'
@@ -9,17 +8,6 @@ import type { GetServerSideProps, NextPage } from 'next'
 type Props = GetServers
 
 const Server: NextPage<Props> = ({ getServers }) => {
-  console.log(getServers)
-  const cellsText = [
-    'Usage',
-    'Hostname',
-    'IP',
-    'Username',
-    'Password',
-    'Device',
-    'Port',
-    'URL'
-  ]
   return (
     <>
       <ServerCellDataContext.Provider value={getServers}>
